added create user option on admin

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->integer('roles_id')->default(3);
+            $table->unsignedBigInteger('roles_id')->default(3);
             $table->string('name');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
